ByteBufferPool factory to private module UndertowModule
